ReplicationLanteriRampini2023

Documentation for ReplicationLanteriRampini2023.

ReplicationLanteriRampini2023.fb_dyn_unc

ReplicationLanteriRampini2023.getDS_shocks_fb

ReplicationLanteriRampini2023.plot_results

ReplicationLanteriRampini2023.rouwen

ReplicationLanteriRampini2023.run_model

ReplicationLanteriRampini2023.fb_dyn_unc — Method

fbdynunc(xx::Array{Float64}, Par::Parameters, Fun::Dict, q::Real, m::Any) Returns an array zz with the first-order conditions for new and old capital at the optimal allocation (see equations 38 and 39 in the paper) that will be used in the policy function iteration. The expression of these FOC is given explicitly as in the Matlab replication files. The inputs correspond to an array with some level of new and old capital (xx), the parameter structure defined before with the associated dictionary with functions at the core of the setup (notably the policy function with its derivatives and production function with its derivatives), a price for capital q and another parameter m. In the next function (getDSshocksfb), m is defined as the matrix of expected payoffs.

ReplicationLanteriRampini2023.getDS_shocks_fb — Method

getDSshocksfb(Ps::Matrix,sgrid, kNfb,kUfb,q0) This function uses NLsolve to solve for the FOC found with fbdynunc ans deduce the level of old capital supplied and demanded optimally by firms of a given cohort. Its arguments are: Ps, the transition matrix of the idiosyncratic shock found usinf Rouwenhorst method; sgrid, theexponential grid of the shock; kNfb and kUfb the levels of new and old capital, and q0 an initial price of capital. It returns a tuple with the demand and the supply of old capital.

ReplicationLanteriRampini2023.plot_results — Method

This function plots the results of the model if one wants.

ReplicationLanteriRampini2023.rouwen — Method

Discretization method for a AR(1) process of the form $z_{t+1} = \mu + \rho z_{t} + u_t$ where $u_t \sim \mathcal{N}(0, \sigma)$ It must be defined as: rouwen(N:: number nodes, $\mu$:: mean of the process, $\sigma$:: sd of the innovation, $\rho$:: persistence). N must be an integer and all other inputs are Floats. The paper considers N=2. Returns a tuple with the discretized grid for the realizations of the shock $Z$ (a vector of size N), and its transition matrix $\Pi$ (size NxN).

ReplicationLanteriRampini2023.run_model — Method

We build a structure with the parameters given as inputs for the model. - A: technology/productivity parameter on the production function - β: discount rate - $\rho$: exit probability - α: curvature of the production function (power on capital stock) - θ: collaterazibility (if θ=0 there is no borrowing in the economy) - δn: depreciation rate of new capital - δu: depreciation rate of old capital - γ: measure of first cohort of firms entering the economy - ρs: persistence idiosyncratic productivity shock s ~ AR(1) - σs: standard deviation innovation of the idiosyncratic productivity shock s - s_grid: grid for the idiosyncratic productivity shock s - Ps: probability transition matrix (will follow form Rouwenhorst method) - w0: initial net worth for new entrants - a: elasticity of output with respect to capital - ϵ: CES elasticity of substitution - χ0: Cost of raising equity parameters - χ1: Cost of raising equity parameters