ReplicationLanteriRampini2023

Documentation for ReplicationLanteriRampini2023.

ReplicationLanteriRampini2023.fb_dyn_unc

ReplicationLanteriRampini2023.getDS_shocks_fb

ReplicationLanteriRampini2023.fb_dyn_unc — Method

fbdynunc(xx::Array{Float64}, Par::Parameters, Fun::Dict, q::Real, m::Any) Returns an array zz with the first-order conditions for new and old capital at the optimal allocation (see equations 38 and 39 in the paper) that will be used in the policy function iteration. The expression of these FOC is given explicitly as in The inputs correspond to an array with some level of new and old capital (xx), the parameter structure defined before with the associated dictionary with functions at the core of the setup (notably the policy function with its derivatives and production function with its derivatives), a price for capital q and another parameter m. In the next function (getDSshocksfb), m is defined as the matrix of expected payoffs.

ReplicationLanteriRampini2023.getDS_shocks_fb — Method

getDSshocksfb(Ps::Matrix,sgrid, kNfb,kUfb,q0) This function uses NLsolve to solve for the FOC found with fbdynunc ans deduce the level of old capital supplied and demanded optimally by firms of a given cohort. Its arguments are: Ps, the transition matrix of the idiosyncratic shock found usinf Rouwenhorst method; sgrid, theexponential grid of the shock; kNfb and kUfb the levels of new and old capital, and q0 an initial price of capital. It returns a tuple with the demand and the supply of old capital.